+++ CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 #### Added
 
 - `generate_random_felt()` for generating (pseudo) random felt value.
-- Printing information about compiling Sierra files by `universal-sierra-compiler` before tests execution
+- Printing information about compiling Sierra files by `universal-sierra-compiler`
 
 ### Cast